add .long()

+++ score_sde/score_fn.py
@@ -20,7 +20,7 @@ class DiffusionEmbedding(nn.Module):
         self.projection2 = nn.Linear(proj_dim, proj_dim)
 
     def forward(self, diffusion_step):
-        x = self.embedding[diffusion_step]
+        x = self.embedding[diffusion_step.long()]
         x = self.projection1(x)
         x = silu(x)
         x = self.projection2(x)
